Use -Wall -Werror and fix HLint suggestions

+++ Main.hs
@@ -2,4 +2,4 @@ module Main where
 import Conway
 
 main :: IO ()
-main = runLife 0 startingCells
+main = runLife (0::Integer) startingCells
